Move inflateHeatmaps to library
