+++ src/js/annotations/heatmap-lib.js
@@ -6,6 +6,15 @@ var reservedTrackKeys = [
   'name', 'start', 'length', 'trackIndex', 'trackIndexOriginal', 'color'
 ];
 
+var defaultHeatmapColors = [
+  ['00B', 'F00'],
+  ['00B', 'DDD', 'F00'],
+  ['00B', 'AAB', 'FAA', 'F00'],
+  ['00B', 'AAB', 'DDD', 'FAA', 'F00'],
+  [], [], [], [], [], [], [], [], [], [], [], // TODO: Use color palette module
+  ['00D', '22D', '44D', '66D', '88D', 'AAD', 'CCD', 'DDD', 'FCC', 'FAA', 'F88', 'F66', 'F44', 'F22', 'F00']
+]
+
 /**
  * Get label text for displayed tracks from annotation container metadata,
  * heatmap keys, or annotation container keys
@@ -35,4 +44,63 @@ function getLabels(ideo) {
   return labels;
 }
 
-export default getLabels;
+/**
+ * Apply heatmap thresholds that are passed in as annotation metadata
+ */
+function inflateThresholds(ideo) {
+  var thresholds, colors, crudeThresholds;
+
+  if (!ideo.rawAnnots.metadata.heatmapThresholds) return;
+
+  thresholds = ideo.rawAnnots.metadata.heatmapThresholds;
+  
+  colors = defaultHeatmapColors[thresholds.length - 1];
+  thresholds = thresholds.map((d, i) => {
+    return [d, '#' + colors[i]];
+  });
+
+  // Coarsen thresholds, emphasize outliers, widen normal range.
+  // TODO: Generalize this for arbitrary number of thresholds.
+  crudeThresholds = [
+    [thresholds[4][0], thresholds[0][1]],
+    [thresholds[6][0], thresholds[3][1]],
+    [thresholds[9][0], thresholds[7][1]],
+    [thresholds[11][0], thresholds[10][1]],
+    [thresholds[14][0], thresholds[14][1]]
+  ]
+
+  thresholds = crudeThresholds;
+
+  thresholds[thresholds.length - 1][0] = '+';
+
+  return thresholds;
+}
+
+/**
+ * Set needed configuration options from raw annotation data.
+ * Simplifies heatmap API by inferring reasonable defaults.
+ */
+function inflateHeatmaps(ideo) {
+  var i, labels, heatmaps, annotationTracks, rawAnnots, displayedTracks,
+    thresholds = ideo.config.heatmapThresholds;
+
+  heatmaps = [];
+  rawAnnots = ideo.rawAnnots;
+  labels = rawAnnots.keys.slice(3,);
+
+  annotationTracks = [];
+  displayedTracks = [];
+  if (rawAnnots.metadata) thresholds = inflateThresholds(ideo);
+
+  for (i = 0; i < labels.length; i++) {
+    heatmaps.push({key: labels[i], thresholds: thresholds});
+    annotationTracks.push({id: labels[i]});
+    displayedTracks.push(i + 1)
+  }
+  ideo.config.annotationsNumTracks = labels.length;
+  ideo.config.annotationsDisplayedTracks = displayedTracks;
+  ideo.config.heatmaps = heatmaps;
+  ideo.config.annotationTracks = annotationTracks;
+}
+
+export {getLabels, inflateHeatmaps};
